Update TCOD::Console methods
The console methods are brought closer to the reference from
python-tcod. This makes a number of the older API methods (in
particular the printing ones) obsolete, although not all of the
old methods are removed in this commit: I'm not sure how aggressive
we can be at the moment.

To add support for a custom clear method for console that takes
optional color and character parameters, this commit makes use of
the bundling mechanic in FFI::Platypus to add some C glue between
the Perl and the C code. The author is impressed with how easy and
painless this was 🎉